requirements.txt
flask_restful
gpiozero
mysql-connector
sshtunnel